+++ src/components/roundabout/components/roundabout-container.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { createCustomizableLunaticField } from '../../commons';
+
+function RoundaboutContainer({ children }) {
+	return <div className="lunatic-roundabout">{children}</div>;
+}
+
+export default createCustomizableLunaticField(
+	RoundaboutContainer,
+	'RoundaboutContainer'
+);
