+++ src/components/roundabout/components/roundabout-it-container.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { createCustomizableLunaticField } from '../../commons';
+
+function RoundaboutItContainer({ children }) {
+	return <div className="roundabout-iteration-title">{children}</div>;
+}
+
+export default createCustomizableLunaticField(
+	RoundaboutItContainer,
+	'RoundaboutItContainer'
+);
