+++ src/components/roundabout/components/roundabout-label.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { createCustomizableLunaticField } from '../../commons';
+
+function RoundaboutLabel({ value }) {
+	return <div className="roundabout-label">{value}</div>;
+}
+
+export default createCustomizableLunaticField(
+	RoundaboutLabel,
+	'RoundaboutLabel'
+);
