+++ src/components/roundabout/components/roundabout-pending.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { createCustomizableLunaticField } from '../../commons';
+
+function RoundaboutPending() {
+	return <div className="roundabout-pending">Pending...</div>;
+}
+
+export default createCustomizableLunaticField(
+	RoundaboutPending,
+	'RoundaboutPending'
+);
